requirements.txt
asgiref==3.8.1
Django==5.0.4
django-colorfield==0.11.0
django-dotenv==1.4.2
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.5.0
tzdata==2024.1
